Extract new pipeline/firehose code into new crate.
