+++ crates/bimm-firehose/Cargo.toml
@@ -1,27 +1,23 @@
 [package]
-name = "bimm"
+name = "bimm-firehose"
 edition.workspace = true
 version.workspace = true
-license.workspace = true
 repository.workspace = true
-documentation = "https://docs.rs/bimm/latest/"
-description = "burn image models"
+license.workspace = true
 
 [lints]
 workspace = true
 
 [dependencies]
 burn = { workspace = true }
-bimm-contracts = { workspace = true }
 serde = { workspace = true, features = ["derive"] }
+serde_json = { workspace = true }
 
 [dev-dependencies]
 burn = { workspace = true, features = ["ndarray"] }
 
-num-traits = { workspace = true }
-hamcrest = { workspace = true }
+indoc = { workspace = true }
 
 [features]
 nightly = []
-burn_0_18_0 = []
