+++ crates/bimm-firehose/src/lib.rs
@@ -1,0 +1,4 @@
+/// New Data Table module.
+pub mod firehose;
+/// New Data Pipeline module.
+pub mod pipeline;
