+++ crates/bimm-firehose/src/pipeline/data_load_error.rs
