+++ crates/bimm-firehose/src/pipeline/mod.rs
