--- crates/bimm/src/data/mod.rs
@@ -1,5 +1,0 @@
-/// New Data Pipeline module.
-pub mod pipeline;
-
-/// New Data Table module.
-pub mod firehose;
